split get-token tests
